+++ src/main/java/book/store/onlinebookstore/service/impl/CategoryServiceImpl.java
@@ -21,7 +21,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public List<CategoryDto> findAll(Pageable pageable) {
-        return categoryRepository.findAll().stream()
+        return categoryRepository.findAll(pageable).stream()
                 .map(categoryMapper::toDto)
                 .toList();
     }
@@ -51,6 +51,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (id < 1) {
+            throw new IndexOutOfBoundsException("Index cannot be less then 1");
+        }
         categoryRepository.deleteById(id);
     }
 }
